requirements.txt
flask
yt-dlp
moviepy
groq
python-dotenv